Move from using double to using float

+++ source/mathlib.c
@@ -1,50 +1,50 @@
 typedef struct Point2DStruct
 {
-    double x;
-    double y;
+    float x;
+    float y;
 }Point2D;
 
 typedef struct Vector3Struct
 {
-    double x;
-    double y;
-    double z;
+    float x;
+    float y;
+    float z;
 }Vector3;
 
 typedef struct QuaternionStruct
 {
-    double x;
-    double y;
-    double z;
-    double w;
+    float x;
+    float y;
+    float z;
+    float w;
 }Quaternion;
 
 typedef struct Matrix4x4Struct
 {
-    double matrix[4][4];
+    float matrix[4][4];
 }Matrix4x4;
 
 const Matrix4x4 emptyMatrix;
 
-Point2D createPoint2D(double x, double y);
+Point2D createPoint2D(float x, float y);
 Point2D project(short width, short height, Vector3 vertex, Matrix4x4 projectionMatrix);
-Vector3 createVector3(double x, double y, double z);
-Vector3 scaleVector3(Vector3 vector, double scale);
+Vector3 createVector3(float x, float y, float z);
+Vector3 scaleVector3(Vector3 vector, float scale);
 Vector3 normalizeVector3(Vector3 vector);
 Vector3 subtractVector3(Vector3 a, Vector3 b);
 Vector3 crossProductVector3(Vector3 a, Vector3 b);
-double dotProductVector3(Vector3 a, Vector3 b);
-double magnitudeVector3(Vector3 vector);
+float dotProductVector3(Vector3 a, Vector3 b);
+float magnitudeVector3(Vector3 vector);
 Vector3 transformVector3ByMatrix(Vector3 vector, Matrix4x4 matrix);
-Quaternion createQuaternion(double x, double y, double z, double w);
-Quaternion vectorToQuaternion(Vector3 vector, double scalar);
+Quaternion createQuaternion(float x, float y, float z, float w);
+Quaternion vectorToQuaternion(Vector3 vector, float scalar);
 Matrix4x4 createLookAtMatrix(Vector3 cameraPosition, Vector3 cameraTarget, Vector3 cameraUp);
-Matrix4x4 createRotationXYZMatrix(double x, double y, double z);
-Matrix4x4 createPerspectiveMatrix(double fov, double aspectRatio, double near, double far);
-Matrix4x4 createTranslationMatrix(double x, double y, double z);
+Matrix4x4 createRotationXYZMatrix(float x, float y, float z);
+Matrix4x4 createPerspectiveMatrix(float fov, float aspectRatio, float near, float far);
+Matrix4x4 createTranslationMatrix(float x, float y, float z);
 Matrix4x4 multiplyMatrices(Matrix4x4 a, Matrix4x4 b);
 
-Point2D createPoint2D(double x, double y)
+Point2D createPoint2D(float x, float y)
 {
     Point2D p;
 
@@ -59,13 +59,13 @@ Point2D project(short width, short height, Vector3 vertex, Matrix4x4 projectionM
     Point2D result;
     Vector3 transformed = transformVector3ByMatrix(vertex, projectionMatrix);
 
-    result.x = transformed.x * width + width / 2.0;
-    result.y = -transformed.y * height + height / 2.0;
+    result.x = transformed.x * width + width / 2.0f;
+    result.y = -transformed.y * height + height / 2.0f;
 
     return result;
 }
 
-Vector3 createVector3(double x, double y, double z)
+Vector3 createVector3(float x, float y, float z)
 {
     Vector3 vector;
 
@@ -76,7 +76,7 @@ Vector3 createVector3(double x, double y, double z)
     return vector;
 }
 
-Vector3 scaleVector3(Vector3 vector, double scale)
+Vector3 scaleVector3(Vector3 vector, float scale)
 {
     Vector3 result;
 
@@ -89,15 +89,15 @@ Vector3 scaleVector3(Vector3 vector, double scale)
 
 Vector3 normalizeVector3(Vector3 vector)
 {
-    double magnitude = magnitudeVector3(vector);
+    float magnitude = magnitudeVector3(vector);
 
-    if (abs(magnitude) <= 0.0001)
+    if (abs(magnitude) <= 0.0001f)
     {
         DEBUG_MSG_FROM("Failed: Vector magnitude was 0.", "normalizeVector3");
-        return createVector3(0.0, 0.0, 0.0);
+        return createVector3(0.0f, 0.0f, 0.0f);
     }
 
-    return scaleVector3(vector, 1.0 / magnitude);
+    return scaleVector3(vector, 1.0f / magnitude);
 }
 
 Vector3 subtractVector3(Vector3 a, Vector3 b)
@@ -116,19 +116,19 @@ Vector3 crossProductVector3(Vector3 a, Vector3 b)
     return vector;
 }
 
-double dotProductVector3(Vector3 a, Vector3 b)
+float dotProductVector3(Vector3 a, Vector3 b)
 {
     return a.x * b.x + a.y * b.y + a.z * b.z;
 }
 
-double magnitudeVector3(Vector3 vector)
+float magnitudeVector3(Vector3 vector)
 {
     return sqrt(dotProductVector3(vector, vector));
 }
 
 Vector3 transformVector3ByMatrix(Vector3 vector, Matrix4x4 matrix)
 {
-    double divisor;
+    float divisor;
     Quaternion quaternion;
     Vector3 result;
 
@@ -142,21 +142,21 @@ Vector3 transformVector3ByMatrix(Vector3 vector, Matrix4x4 matrix)
     divisor = matrix.matrix[0][3] * vector.x + matrix.matrix[1][3] * vector.y +
               matrix.matrix[2][3] * vector.z + matrix.matrix[3][3];
 
-    if (abs(divisor) < 0.0001)
+    if (abs(divisor) < 0.0001f)
     {
-        divisor = 0.0001;
+        divisor = 0.0001f;
         DEBUG_MSG_FROM("Failed: Can't divide by 0. CameraPosition and Vertex are equal!",
                        "transformVector3ByMatrix");
     }
 
-    quaternion.w = 1.0 / divisor;
+    quaternion.w = 1.0f / divisor;
 
     result = createVector3(quaternion.x * quaternion.w, quaternion.y * quaternion.w, quaternion.z * quaternion.w);
 
     return result;
 }
 
-Quaternion createQuaternion(double x, double y, double z, double w)
+Quaternion createQuaternion(float x, float y, float z, float w)
 {
     Quaternion q;
 
@@ -168,7 +168,7 @@ Quaternion createQuaternion(double x, double y, double z, double w)
     return q;
 }
 
-Quaternion vectorToQuaternion(Vector3 vector, double scalar)
+Quaternion vectorToQuaternion(Vector3 vector, float scalar)
 {
     Quaternion q;
 
@@ -191,33 +191,33 @@ Matrix4x4 createLookAtMatrix(Vector3 cameraPosition, Vector3 cameraTarget, Vecto
     result.matrix[0][0] =  xAxis.x;
     result.matrix[0][1] =  yAxis.x;
     result.matrix[0][2] =  zAxis.x;
-    result.matrix[0][3] =  0.0;
+    result.matrix[0][3] =  0.0f;
 
     result.matrix[1][0] =  xAxis.y;
     result.matrix[1][1] =  yAxis.y;
     result.matrix[1][2] =  zAxis.y;
-    result.matrix[1][3] =  0.0;
+    result.matrix[1][3] =  0.0f;
 
     result.matrix[2][0] =  xAxis.z;
     result.matrix[2][1] =  yAxis.z;
     result.matrix[2][2] =  zAxis.z;
-    result.matrix[2][3] =  0.0;
+    result.matrix[2][3] =  0.0f;
 
     result.matrix[3][0] = -dotProductVector3(xAxis, cameraPosition);
     result.matrix[3][1] = -dotProductVector3(yAxis, cameraPosition);
     result.matrix[3][2] = -dotProductVector3(zAxis, cameraPosition);
-    result.matrix[3][3] =  1.0;
+    result.matrix[3][3] =  1.0f;
 
     return result;
 }
 
-Matrix4x4 createRotationXYZMatrix(double x, double y, double z)
+Matrix4x4 createRotationXYZMatrix(float x, float y, float z)
 {
     Matrix4x4 result;
 
-    double xcos, ycos, zcos;
-    double xsin, ysin, zsin;
-    double zcosxsin, xcoszcos, ysinzsin;
+    float xcos, ycos, zcos;
+    float xsin, ysin, zsin;
+    float zcosxsin, xcoszcos, ysinzsin;
 
     xcos = cos(x);
     xsin = sin(x);
@@ -233,48 +233,48 @@ Matrix4x4 createRotationXYZMatrix(double x, double y, double z)
     result.matrix[0][0] =  ycos * zcos;
     result.matrix[0][1] =  zcosxsin * ysin + xcos * zsin;
     result.matrix[0][2] =  -xcoszcos * ysin + xsin * zsin;
-    result.matrix[0][3] =  0.0;
+    result.matrix[0][3] =  0.0f;
 
     result.matrix[1][0] =  -ycos * zsin;
     result.matrix[1][1] =  xcoszcos - xsin * ysinzsin;
     result.matrix[1][2] =  zcosxsin + xcos * ysinzsin;
-    result.matrix[1][3] =  0.0;
+    result.matrix[1][3] =  0.0f;
 
     result.matrix[2][0] =  ysin;
     result.matrix[2][1] =  -ycos * xsin;
     result.matrix[2][2] =  xcos * ycos;
-    result.matrix[2][3] =  0.0;
+    result.matrix[2][3] =  0.0f;
 
-    result.matrix[3][0] =  0.0;
-    result.matrix[3][1] =  0.0;
-    result.matrix[3][2] =  0.0;
-    result.matrix[3][3] =  1.0;
+    result.matrix[3][0] =  0.0f;
+    result.matrix[3][1] =  0.0f;
+    result.matrix[3][2] =  0.0f;
+    result.matrix[3][3] =  1.0f;
 
     return result;
 }
 
-Matrix4x4 createPerspectiveMatrix(double fov, double aspectRatio, double near, double far)
+Matrix4x4 createPerspectiveMatrix(float fov, float aspectRatio, float near, float far)
 {
-    double yScale = 1.0 / tan(fov * 0.5);
-    double xScale = yScale / aspectRatio;
+    float yScale = 1.0f / tan(fov * 0.5f);
+    float xScale = yScale / aspectRatio;
 
     Matrix4x4 result;
 
-    if (fov <= 0.0 || fov >= PI)
+    if (fov <= 0.0f || fov >= PI)
     {
         char temp[32];
         sprintf(temp, "Failed: Invalid fov: %f.", fov);
         DEBUG_MSG_FROM(temp, "createPerspectiveMatrix");
         return emptyMatrix;
     }
-    if (near <= 0.0)
+    if (near <= 0.0f)
     {
         char temp[32];
         sprintf(temp, "Failed: Invalid near: %f.", near);
         DEBUG_MSG_FROM(temp, "createPerspectiveMatrix");
         return emptyMatrix;
     }
-    if (far <= 0.0)
+    if (far <= 0.0f)
     {
         char temp[32];
         sprintf(temp, "Failed: Invalid far: %f.", far);
@@ -290,44 +290,44 @@ Matrix4x4 createPerspectiveMatrix(double fov, double aspectRatio, double near, d
     }
 
     result.matrix[0][0] = xScale;
-    result.matrix[0][1] = result.matrix[0][2] = result.matrix[0][3] = 0.0;
+    result.matrix[0][1] = result.matrix[0][2] = result.matrix[0][3] = 0.0f;
 
     result.matrix[1][1] = yScale;
-    result.matrix[1][0] = result.matrix[1][2] = result.matrix[1][3] = 0.0;
+    result.matrix[1][0] = result.matrix[1][2] = result.matrix[1][3] = 0.0f;
 
     result.matrix[2][2] = far / (near - far);
-    result.matrix[2][3] = -1.0;
-    result.matrix[2][0] = result.matrix[2][1] = 0.0;
+    result.matrix[2][3] = -1.0f;
+    result.matrix[2][0] = result.matrix[2][1] = 0.0f;
 
     result.matrix[3][2] = near * far / (near - far);
-    result.matrix[3][0] = result.matrix[3][1] = result.matrix[3][3] = 0.0;
+    result.matrix[3][0] = result.matrix[3][1] = result.matrix[3][3] = 0.0f;
 
     return result;
 }
 
-Matrix4x4 createTranslationMatrix(double x, double y, double z)
+Matrix4x4 createTranslationMatrix(float x, float y, float z)
 {
     Matrix4x4 result;
 
-    result.matrix[0][0] = 1.0;
-    result.matrix[0][1] = 0.0;
-    result.matrix[0][2] = 0.0;
-    result.matrix[0][3] = 0.0;
+    result.matrix[0][0] = 1.0f;
+    result.matrix[0][1] = 0.0f;
+    result.matrix[0][2] = 0.0f;
+    result.matrix[0][3] = 0.0f;
 
-    result.matrix[1][0] = 0.0;
-    result.matrix[1][1] = 1.0;
-    result.matrix[1][2] = 0.0;
-    result.matrix[1][3] = 0.0;
+    result.matrix[1][0] = 0.0f;
+    result.matrix[1][1] = 1.0f;
+    result.matrix[1][2] = 0.0f;
+    result.matrix[1][3] = 0.0f;
 
-    result.matrix[2][0] = 0.0;
-    result.matrix[2][1] = 0.0;
-    result.matrix[2][2] = 1.0;
-    result.matrix[2][3] = 0.0;
+    result.matrix[2][0] = 0.0f;
+    result.matrix[2][1] = 0.0f;
+    result.matrix[2][2] = 1.0f;
+    result.matrix[2][3] = 0.0f;
 
     result.matrix[3][0] = x;
     result.matrix[3][1] = y;
     result.matrix[3][2] = z;
-    result.matrix[3][3] = 1.0;
+    result.matrix[3][3] = 1.0f;
 
     return result;
 }
